Implemented the users activity badge system
If the users was active in previous 25 hours (25 not 24, beacause im
flooring time, so 1 hr means that user was active <1-2) hrs ago), there
is grey badge below their profile image. If they're active, there is a
green circle instead.

Every minute the app send GET requests to the server to get updates on
the users activities.

I also rearranged a little bit App, Chat and ActiveChat structure to
make them more sensefully, also now the STOMP connection is held in the
App component, so it's not being occured if users goes for example to
/profile.