Improved look of the app and structure
Fixed Chats page look, so now there is not excess Y scrollbar and the
page is still nice centered.

Restructered the App component. Now it doesn't handle WebSocket
conneciton. It is done in AppWithNavbarAndConnection. This way, the
conneciton is not present in Login and Register which was a bug.
Also cleared Login and Register components a little bit.

Little look improvement for Login and Register, fixed wrong validity condition
for username in Register.